requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pyspellchecker>=0.8.3
html2text>=2020.1.16
lxml>=4.9.0
pyyaml>=6.0.0
tqdm>=4.66.0